Add detector
A base class for `Detector` and its result (`Detection`) is introduce.

An naive example of `Detector` based on "Hough Circle Detection" is
added to show case the usage of the newly introduced classes and their
interfaces.

The `visualization` and `dataset` module are refactored to support
visualization of the detection result in two cases:
* when working with a `dataset_manager` instance which handles
  everything regarding data (images, annotations, ...). This is mostly
  useful for development.
* when deploying detector directly on some arbitrary input data, where
  no dataset structure or annotation is available.

Furthermore:
* data structures such as `Frame` and `BoundingBox` are refactored and
  moved to a separate module.

Signed-off-by: Saeed Gholami Shahbandi <saeed.gh.sh@gmail.com>