Refactor Frame to require id and not Optional
When the detector entry point is called in direct mode, there was no
frame id and the output files where given 'None_None_<somthing>.png`
which would be overwriting multiple output on top of each other.

To remedy this, `load_images` (used by detector entry point instead of
`DatasetManager` in `direct` mode) is changed to return `dict` instead
of `list` when the keys are based on image path and used as `frame_id`
in instantiating `Frame` objects.

Signed-off-by: Saeed Gholami Shahbandi <saeed.gh.sh@gmail.com>

 #	libs/detection/hog_svm_detector.py