Refactor dataset splitting functions
The new version is more generic, and doesn't assume what the subsets
resulting from split are, as former version assumed "train" and "test"
would be the subsets. Avoiding that assumption would allow using it
also to split train set into train and validation for the training and
regularization of classifiers.

Signed-off-by: Saeed Gholami Shahbandi <saeed.gh.sh@gmail.com>